Branch update to include CCweather_version files
This is an update to include the model version used in the simulation for the "Near-, medium- and long-term impacts of climate change on the thermal energy consumption of buildings in Finland under RCP climate scenarios" article. 

Not to be merged with the main "SBuM" branch. Separate merge will be conducted later.